linter/requirements.txt
pyyaml ~= 6.0.3
schema ~= 0.7.8
